Howl's Moving Castle Project Using HTML and CSS
Link to webpage:
https://tamegomes.github.io/Howl-s-Moving-Castle/

This webpage allows the user to see the sinopsis of Howl's Moving Castle movie, to watch the movie on Netflix if the user has a account and to watch the Trailer on Youtube. The user also can acess the social medias of Studio Ghibli by clicking on the respective icons.